Root block alloc in basegrid now distributed
Previously, root block allocation was done only by the master, and later
load balanced. This is problematic when the number of root blocks is
large and does not fit into the master's memory. Root block allocation
is now done in parallel by all ranks, each rank allocating only a
fraction of all root blocks. This is done "naively" (not following the
Hilbert curve), but after ICs are imposed a loadBalance() is done.